Fix cuda installs (#652)
* Correct cuda_samples_path for secure nfs-homedirs

The ``rocky`` user's homedir has moved from ``/home/rocky`` to
``/var/lib/rocky``. This updates ``cuda_samples_path`` to use that.

* Hardcode cuda_version_short and remove lookup

We already hardcode the version of cuda which we install. This ensures
that we will also use the requested version when multiple cuda versions
are installed.

Moves ``cuda_version_short`` to be next to ``cuda_package_version`` so
it's harder to miss updating one when the other is changed.

* Install cmake and cuda-toolkit