Use checksum verification for CernVM-FS GPG key (#641)
The cvmrepo repository is sometimes down. This avoids fetching the GPG
key each time if it was already done, preventing Ansible failures.

It also verifies that the key is the expected one instead of blindly
trusting any GPG key.

This should not require much maintenance since the key appears to be the
same since it was generated in 2010.

Co-authored-by: Steve Brasier <33413598+sjpb@users.noreply.github.com>