Enable persistent journaling
In the cloud images, /var/log/journal doesn't exist. When this
directory does not exist, journald will not write any files to
disk, and as such, logging data is lost accross reboots.

Persistent logging is useful for debugging issues that caused the nodes
to reboot (provided that the logs are flushed to disk).