Reformat compute_fixed_image user_data (#847)
When the additional_cloud_config support was added, newline
characters were also added outside of the conditional block. This
means compute_fixed_image instances would be shceduled for
replacement on the next tofu apply. This breaks slurm controlled
rebuilds, as these fixed nodes should not be replaced at this step.

Correct this by moving the newlines to be within the conditional
block. The double newline is presevred to avoid breaking systems
which are already using both additional_cloud_config and slurm
controlled rebuild.